Refine spec file

+++ SPECS/cozy-stack.spec
@@ -3,7 +3,7 @@
 
 Name:           cozy-stack
 Version:        2018M1S6
-Release:        0%{?dist}
+Release:        1%{?dist}
 Summary:        Cozy: Simple, Versatile, Yours
 
 Group:          Applications/System
@@ -24,7 +24,7 @@ in the same private space.
 This package installs the cozy stack.
 
 %prep
-%autosetup -n cozy-stack
+%autosetup -n %{name}
 
 %build
 export GOPATH="$(pwd)"
@@ -33,14 +33,14 @@ cd src/%{repo}
 COZY_ENV=production GOOS=linux GOARCH=amd64 VERSION_STRING=prod ./scripts/build.sh install
 
 %install
-mkdir -p %{buildroot}/usr/local/bin
-cp bin/cozy-stack %{buildroot}/usr/local/bin/cozy-stack
+mkdir -p %{buildroot}%{_bindir}
+cp bin/%{name} %{buildroot}%{_bindir}
 mkdir -p %{buildroot}%{_datarootdir}/cozy
 cp src/%{repo}/scripts/konnector-node-run.sh %{buildroot}%{_datarootdir}/cozy
 cp src/%{repo}/scripts/konnector-nsjail-run.sh %{buildroot}%{_datarootdir}/cozy
 cp src/%{repo}/scripts/konnector-rkt-run.sh %{buildroot}%{_datarootdir}/cozy
 mkdir -p %{buildroot}%{_unitdir}
-cat << EOF > %{buildroot}%{_unitdir}/cozy-stack.service
+cat << EOF > %{buildroot}%{_unitdir}/%{name}.service
 [Unit]
 Description=Cozy service
 Wants=couchdb.service
@@ -50,7 +50,7 @@ After=network.target couchdb.service
 User=cozy
 Group=cozy
 PermissionsStartOnly=true
-ExecStart=/usr/local/bin/cozy-stack serve
+ExecStart=/usr/bin/cozy-stack serve
 Restart=always
 
 [Install]
@@ -58,42 +58,56 @@ WantedBy=multi-user.target
 EOF
 mkdir -p %{buildroot}%{_sysconfdir}/cozy
 cp %{SOURCE1} %{buildroot}%{_sysconfdir}/cozy/cozy.yaml
-sed -i 's@# url: file://localhost/var/lib/cozy@url: file://localhost/var/lib/cozy@;s@/path/to/key.enc@/home/cozy/keys/key.enc@;s@/path/to/key.dec@/home/cozy/keys/key.dec@;s@cmd: ./scripts@cmd: %{_datarootdir}/cozy@g' %{buildroot}%{_sysconfdir}/cozy/cozy.yaml
+sed -i 's@# url: file://localhost/var/lib/cozy@url: file://localhost/var/lib/cozy@;s@/path/to/key.enc@/var/lib/cozy/keys/key.enc@;s@/path/to/key.dec@/var/lib/cozy/keys/key.dec@;s@cmd: ./scripts@cmd: %{_datarootdir}/cozy@g' %{buildroot}%{_sysconfdir}/cozy/cozy.yaml
 
 %pre
 
 %post
-useradd cozy
-chown cozy: -R /etc/cozy
-mkdir -p /var/lib/cozy && chown cozy: /var/lib/cozy
-mkdir -p /var/log/cozy && chown cozy: /var/log/cozy
-mkdir -p /home/cozy/keys && chown cozy: /home/cozy/keys
-su - cozy -c "cozy-stack config gen-keys ~/keys/key &> /dev/null"
+if [ "$1" = 1 ]; then
+	groupadd --system cozy
+	useradd --system --gid cozy -d /var/lib/cozy -s /sbin/nologin cozy
+	chown cozy: -R /etc/cozy
+	mkdir -p /var/lib/cozy/keys && chown cozy: -R /var/lib/cozy
+	chmod 700 /var/lib/cozy/keys
+	mkdir -p /var/log/cozy && chown cozy: /var/log/cozy
+	su - cozy -s /bin/bash -c "cozy-stack config gen-keys /var/lib/cozy/keys/key &> /dev/null"
+fi
 
+if [ "$1" = 2 ]; then
+    systemctl daemon-reload
+    if systemctl is-enabled cozy-stack &>/dev/null || systemctl is-active cozy-stack &>/dev/null ; then
+        systemctl restart cozy-stack
+    fi
+fi
 
 %preun
 if [ "$1" = 0 ]; then 
-    systemctl stop cozy-stack
-    systemctl disable cozy-stack
+    if systemctl is-enabled cozy-stack &>/dev/null || systemctl is-active cozy-stack &>/dev/null ; then
+    	systemctl stop cozy-stack
+    	systemctl disable cozy-stack
+	fi
 fi
 
 %postun
 if [ "$1" = 0 ]; then
-    mv /home/cozy /home/cozy.bkp
     mv /var/lib/cozy /var/lib/cozy.bkp
-    mv /var/log/cozy /var/log/cozy.bkp
-    userdel cozy
+    userdel -f cozy
+    systemctl daemon-reload
 fi
 
 %files
-/usr/local/bin/cozy-stack
-%{_unitdir}/cozy-stack.service
-%config %{_sysconfdir}/cozy/cozy.yaml
+%{_bindir}/%{name}
+%{_unitdir}/%{name}.service
+%config(noreplace) %{_sysconfdir}/cozy/cozy.yaml
 %{_datarootdir}/cozy/konnector-node-run.sh
 %{_datarootdir}/cozy/konnector-nsjail-run.sh
 %{_datarootdir}/cozy/konnector-rkt-run.sh
 %license src/%{repo}/LICENSE
 
 %changelog
+* Sun Mar 25 2018 Pierre-Alain TORET <pierre-alain.toret@protonmail.com> 2018M1S6-1
+- Improve scriptlets
+- Move things to more standard paths
+
 * Sun Mar 25 2018 Pierre-Alain TORET <pierre-alain.toret@protonmail.com> 2018M1S6-0
 - Inital rpm for release 2018M1S6
